Introduce PartialFile helper
Previously, we created files with their final name and wrote to them.
This will abstract this away a bit, and first create a hidden temporary
file. (In the same directory, so it's on the same file system). When you
are done writing to it, call `finish` and it'll be renamed to the name you
wanted. As a bonus, this also uses a BufWriter so you don't have to.